internal/handler/handler_test.go
package handler

import (
	sqlmock "github.com/DATA-DOG/go-sqlmock"
	"github.com/gin-gonic/gin"
	"net/http"
	"net/http/httptest"
	"strings"
	"testing"
)

func TestCreateInvoice(t *testing.T) {
	// Create a mock database connection
	db, mock, err := sqlmock.New()
	if err != nil {
		t.Fatal(err)
	}
	defer db.Close()

	// Create a mock Gin context with request data
	requestJSON := `{"some": "data"}` // Replace with your test data
	req, err := http.NewRequest("POST", "/create-invoice", strings.NewReader(requestJSON))
	if err != nil {
		t.Fatal(err)
	}
	req.Header.Set("Content-Type", "application/json")
	w := httptest.NewRecorder()
	c, _ := gin.CreateTestContext(w)
	c.Request = req

	// Define expectations for the database mock
	mock.ExpectExec("SELECT create_invoice").WithArgs( /* expected arguments */ ).WillReturnResult(sqlmock.NewResult(1, 1))

	// Call the handler function with the mock database
	CreateInvoice(c, db)

	// Check the response status code and body as needed
	if w.Code != http.StatusOK {
		t.Errorf("Expected status code %d, but got %d", http.StatusOK, w.Code)
	}
	// Check other assertions as needed
}

func TestWithdrawFunds(t *testing.T) {
	// Create a mock database connection
	db, mock, err := sqlmock.New()
	if err != nil {
		t.Fatal(err)
	}
	defer db.Close()

	// Create a mock Gin context with request data
	requestJSON := `{"some": "data"}` // Replace with your test data
	req, err := http.NewRequest("POST", "/withdraw-funds", strings.NewReader(requestJSON))
	if err != nil {
		t.Fatal(err)
	}
	req.Header.Set("Content-Type", "application/json")
	w := httptest.NewRecorder()
	c, _ := gin.CreateTestContext(w)
	c.Request = req

	// Define expectations for the database mock
	mock.ExpectExec("SELECT withdraw_funds").WithArgs( /* expected arguments */ ).WillReturnResult(sqlmock.NewResult(1, 1))

	// Call the handler function with the mock database
	WithdrawFunds(c, db)

	// Check the response status code and body as needed
	if w.Code != http.StatusOK {
		t.Errorf("Expected status code %d, but got %d", http.StatusOK, w.Code)
	}
	// Check other assertions as needed
}
